fixing warning refs #12
git-svn-id: file:///home/behr_mi/git/sumo_synched/trunk@14388 afbd958f-9f77-42d5-a016-97a22340ccf4

+++ sumo/src/microsim/MSVehicle.cpp
@@ -2022,7 +2022,7 @@ MSVehicle::saveState(OutputDevice& out) {
     // here starts the vehicle internal part (see loading)
     std::vector<int> internals;
     internals.push_back(myDeparture);
-    internals.push_back(distance(myRoute->begin(), myCurrEdge));
+    internals.push_back((int)distance(myRoute->begin(), myCurrEdge));
     out.writeAttr(SUMO_ATTR_STATE, toString(internals));
     out.writeAttr(SUMO_ATTR_POSITION, myState.myPos);
     out.writeAttr(SUMO_ATTR_SPEED, myState.mySpeed);
